+++ tests/http.writeEvent.js
@@ -16,4 +16,16 @@ describe('Http Client - Write Event', function() {
             });
         });
     });
+
+    it('Should fail promise if no event data provided', function() {
+        var client = eventstore.http(httpConfig);
+
+        var testStream = 'TestStream-' + uuid.v4();
+        return client.writeEvent(testStream, 'TestEventType'
+            ).then(function() {
+            assert.fail('write should not have succeeded');
+        }).catch(function(err) {
+            assert(err,'error should have been returned');
+        });
+    });
 });